Cleaning up the XorBoard Bench: 5355809
