+++ src/core/board/xorboard/views.rs
@@ -7,28 +7,13 @@
 
 use core::arch::x86_64::*;
 
-use super::{NOWHERE, PieceId, XorBoard, class_index, color_slots, slots};
+use super::{PieceId, XorBoard, class_index, slots};
 use crate::core::{
-    board::bitboard::{atk_bishop, atk_rook, between_bb, line_bb},
+    board::bitboard::line_bb,
     defs::{Bitboard, Color, PieceType, Square},
 };
 #[cfg(not(target_feature = "avx512vpopcntdq"))]
 use crate::weave::Vu64x4;
-
-/// Four lanes of all-ones or all-zeros, indexed by a nibble of the slot mask.
-static LANE_MASK: [[u64; 4]; 16] = {
-    let mut table = [[0u64; 4]; 16];
-    let mut nibble = 0;
-    while nibble < 16 {
-        let mut lane = 0;
-        while lane < 4 {
-            table[nibble][lane] = if nibble >> lane & 1 == 1 { u64::MAX } else { 0 };
-            lane += 1;
-        }
-        nibble += 1;
-    }
-    table
-};
 
 impl XorBoard {
     /// Every square `color` attacks.
@@ -59,58 +44,9 @@ impl XorBoard {
         }
     }
 
-    /// What `id` attacks through a slider standing in its way.
-    ///
-    /// The phantom-piece rule says an x-ray needs an occupancy that does not
-    /// exist, and it holds for every blocker but this one: a slider's own row
-    /// already carries the continuation, so the answer is a load and an AND
-    /// instead of a probe. Only the far segment survives the mask, because the
-    /// stretch back toward `id` is `between_bb` and nothing past `id` is in the
-    /// blocker's row, `id` being what stops that ray.
-    ///
-    /// The blocker has to run along this line for its row to say anything about
-    /// it, so this answers batteries and not x-rays in general: a bishop in a
-    /// rook's way hides whatever is behind it, and no test is needed to drop it
-    /// because a bishop's row meets a rank nowhere.
-    pub fn xray_row(&self, id: PieceId, slider_squares: Bitboard) -> Bitboard {
-        let from = Square(self.squares[id.index()]);
-        let mut through = Bitboard(0);
-        for square in Bitboard(self.rows[id.index()]) & slider_squares {
-            let Some(blocker) = self.id_at(square) else { continue };
-            let past = line_bb(from, square) & !between_bb(from, square) & !from.bitboard() & !square.bitboard();
-            through |= Bitboard(self.rows[blocker.index()]) & past;
-        }
-        through
-    }
-
-    /// Every square the given class attacks.
-    #[inline(always)]
-    pub fn class_attacks(&self, piece: PieceType, color: Color) -> Bitboard {
-        self.union(self.class[class_index(piece, color)])
-    }
-
-    /// Every piece of `color` bar the king, each with the squares it may legally
-    /// use, pins applied by `pinned_row`.
-    #[inline(always)]
-    pub fn legal_rows(&self, color: Color, pinned: Bitboard, ksq: Square) -> impl Iterator<Item = (PieceId, Bitboard)> + '_ {
-        let king = self.class[class_index(PieceType::King, color)];
-
-        slots(color_slots(color) & !king).filter_map(move |id| {
-            let raw = self.squares[id.index()];
-            if raw == NOWHERE {
-                return None;
-            }
-
-            let square = Square(raw);
-            let row = self.row(id);
-            Some((id, if pinned.check_bit(square) { self.pinned_row(id, square, row, ksq) } else { row }))
-        })
-    }
-
-    /// The eval's attack map for `color`: the union of what its pieces may use.
-    #[inline(always)]
-    pub fn attack_map(&self, color: Color, pinned: Bitboard, ksq: Square) -> Bitboard {
-        self.legal_rows(color, pinned, ksq).fold(Bitboard(0), |acc, (_, row)| acc | row)
+    /// Scalar: the only caller is the debug-assert oracle.
+    pub(super) fn class_attacks(&self, piece: PieceType, color: Color) -> Bitboard {
+        slots(self.class[class_index(piece, color)]).fold(Bitboard(0), |acc, id| acc | self.row(id))
     }
 
     /// Mobility counted per piece rather than over the union.
@@ -149,7 +85,7 @@ impl XorBoard {
     /// the tensor does today. Crediting either of the first two with more is
     /// mobility for a move that would leave the king in check.
     #[inline(always)]
-    fn pinned_row(&self, id: PieceId, square: Square, row: Bitboard, ksq: Square) -> Bitboard {
+    pub(super) fn pinned_row(&self, id: PieceId, square: Square, row: Bitboard, ksq: Square) -> Bitboard {
         match self.kind[id.index()] {
             PieceType::Knight => Bitboard(0),
             PieceType::Bishop | PieceType::Rook | PieceType::Queen => row & line_bb(ksq, square),
@@ -186,70 +122,6 @@ impl XorBoard {
                 let folded = _mm_add_epi64(_mm256_castsi256_si128(acc), _mm256_extracti128_si256::<1>(acc));
                 (_mm_extract_epi64::<0>(folded) + _mm_extract_epi64::<1>(folded)) as i32
             }
-        }
-    }
-
-    /// Attacks that pass through exactly one friendly piece, orthogonal and
-    /// diagonal kept apart because the eval scores them apart.
-    ///
-    /// Lifting a slider's own first blockers out of occupancy and probing again
-    /// continues each ray from where it stopped, which is what the tensor's
-    /// second flood-fill does by feeding the friendly-hit squares back in as
-    /// generators. Pinned pieces contribute nothing, matching the tensor.
-    #[inline(always)]
-    pub fn xray_maps(&self, color: Color, pinned: Bitboard, own: Bitboard, occ: Bitboard) -> (Bitboard, Bitboard) {
-        let (mut ortho, mut diag) = (Bitboard(0), Bitboard(0));
-        let (mut ortho_direct, mut diag_direct) = (Bitboard(0), Bitboard(0));
-
-        let mut rest = self.slider_slots & color_slots(color);
-
-        while rest != 0 {
-            let id = PieceId(rest.trailing_zeros() as u8);
-            rest &= rest - 1;
-            // A captured piece keeps its class bit; only its row, square and
-            // mailbox entry are cleared, so liveness has to be tested here.
-            if self.squares[id.index()] == NOWHERE {
-                continue;
-            }
-
-            let from = Square(self.squares[id.index()]);
-            if pinned.check_bit(from) {
-                continue;
-            }
-
-            let kind = self.kind[id.index()];
-            let behind = occ & !(self.row(id) & own);
-
-            if kind != PieceType::Bishop {
-                ortho_direct |= atk_rook(from, occ);
-                ortho |= atk_rook(from, behind);
-            }
-            if kind != PieceType::Rook {
-                diag_direct |= atk_bishop(from, occ);
-                diag |= atk_bishop(from, behind);
-            }
-        }
-        (ortho & !ortho_direct, diag & !diag_direct)
-    }
-
-    /// Reduce-OR over the selected slots. Masked rather than iterated: the
-    /// callers pass sixteen-bit selections, and a lane mask off a nibble table
-    /// beats walking the set bits at that density.
-    #[inline(always)]
-    fn union(&self, slots: u64) -> Bitboard {
-        // SAFETY: AVX2 per the compile_error gate in weave/mod.rs. Each load
-        // covers one group of four of a 32-element array, and the nibble table
-        // is indexed by four bits so it stays inside its sixteen rows.
-        unsafe {
-            let mut acc = _mm256_setzero_si256();
-            for group in 0..8 {
-                let rows = _mm256_loadu_si256(self.rows.as_ptr().add(group * 4).cast());
-                let keep = _mm256_loadu_si256(LANE_MASK.as_ptr().add((slots >> (group * 4)) as usize & 15).cast());
-                acc = _mm256_or_si256(acc, _mm256_and_si256(rows, keep));
-            }
-
-            let folded = _mm_or_si128(_mm256_castsi256_si128(acc), _mm256_extracti128_si256::<1>(acc));
-            Bitboard((_mm_extract_epi64::<0>(folded) | _mm_extract_epi64::<1>(folded)).cast_unsigned())
         }
     }
 }
